day02: finally solved part 2
struggled for way too long with how to do the "forward" action given
a 3array and the delta on the stack. trying to be like "yeah just do
stuff to the first index, then the second index, with this one value on
the top of stack" was just beyond me. finally realized i could just
create a couple more delta vectors and v+ twice in a row, which dodged
the problem.

probably i just failed to create the relevant words to make it easy, but
i definitely felt like i was going against the grain of "how you should
do things" in the language. :|