day11: got silver with complete rewrite
trying to dump a ton of code does not work.
breaking it down into smaller words does.

the big breakthrough that reduced stack woes was figuring out how to
track the already-flashed info within the matrix. this is done by
marking the just-flashed points with an F, applying their bonus to their
neighbors, then clearing the Fs to 0s before repeating.