day12: got gold
coming back after a break helped.

using the REPL to step through stuff by hand, and then canonizing the
working stuff as tests, helped.

the solution is ugly because i just duplicated stuff, rather than
passing in a decision around how to filter things, but i just didn't
want to fight the stack with that. :\

factor is a very nice environment, but i just don't feel like i'm
meshing well with the stack. :(